+++ src/AzzyBot-Next/Utilities/Helpers/AzuraApiEndpoints.cs
@@ -7,6 +7,8 @@ public static class AzuraApiEndpoints
     public const string Clear = "clear";
     public const string Files = "files";
     public const string History = "history";
+    public const string Log = "log";
+    public const string Logs = "logs";
     public const string NowPlaying = "nowplaying";
     public const string Queue = "queue";
     public const string Playlist = "playlist";